+++ quarry/types/buffer/v1_7.py
@@ -1,5 +1,7 @@
 import json
+import string
 import struct
+import sys
 import zlib
 
 from quarry.types.buffer import BufferUnderrun
@@ -7,6 +9,7 @@ from quarry.types.uuid import UUID
 from quarry.types.chunk import BlockArray, LightArray
 from quarry.types.block import OpaqueBlockMap
 
+PY3 = sys.version_info > (3,)
 
 # Python 3 compat
 try:
@@ -74,6 +77,38 @@ class Buffer1_7(object):
             self.pos += length
 
         return data
+
+    def hexdump(self):
+        printable = string.letters + string.digits + string.punctuation
+        data = self.buff[self.pos:]
+        lines = ['']
+        bytes_read = 0
+        while len(data) > 0:
+            data_line, data = data[:16], data[16:]
+
+            l_hex = []
+            l_str = []
+            for i, c in enumerate(data_line):
+                if PY3:
+                    l_hex.append("%02x" % c)
+                    c_str = data_line[i:i + 1]
+                    l_str.append(
+                        c_str if c_str in printable else ".")
+                else:
+                    l_hex.append("%02x" % ord(c))
+                    l_str.append(c if c in printable else ".")
+
+            l_hex.extend(['  '] * (16 - len(l_hex)))
+            l_hex.insert(8, '')
+
+            lines.append("%08x  %s  |%s|" % (
+                bytes_read,
+                " ".join(l_hex),
+                "".join(l_str)))
+
+            bytes_read += len(data_line)
+
+        return "\n    ".join(lines + ["%08x" % bytes_read])
 
     # Basic data types --------------------------------------------------------